+++ Tests/Codeception/_support/Page/PayPalLogin.php
@@ -36,14 +36,23 @@ class PayPalLogin extends Page
 
     public $spinner = '#spinner';
 
+    public $gdprContainer = "#gdpr-container";
+    public $gdprCookieBanner = "#gdprCookieBanner";
     public $acceptAllPaypalCookies = "#acceptAllButton";
 
     public $loginSection = "#loginSection";
+    public $oldLoginSection = "#passwordSection";
 
     public $cancelLink = "#cancelLink";
     public $returnToShop = "#cancel_return";
 
     public $breadCrumb = "#breadCrumb";
+
+    public $paymentConfirmButton = "#payment-submit-btn";
+    public $globalSpinner = "//div[@data-testid='global-spinner']";
+    public $preloaderSpinner = "//div[@id='preloaderSpinner']";
+
+    public $paypalBannerContainer = "//div[@id='paypal-installment-banner-container']";
 
     /**
      * @param string $userName
@@ -101,6 +110,101 @@ class PayPalLogin extends Page
         return new OrderCheckout($I);
     }
 
+    /**
+     * @param string $userName
+     * @param string $userPassword
+     *
+     * @return OrderCheckout
+     */
+    public function checkoutWithStandardPayPal(string $userName, string $userPassword): OrderCheckout
+    {
+        $I = $this->user;
+
+        $this->loginToPayPal($userName, $userPassword);
+
+        $this->confirmPayPal();
+
+        //retry
+        $this->waitForSpinnerDisappearance();
+        $this->confirmPayPal();
+
+        return new OrderCheckout($I);
+    }
+
+    public function loginToPayPal(string $userName, string $userPassword): void
+    {
+        $I = $this->user;
+
+        $this->waitForPayPalPage();
+        $this->removeCookieConsent();
+
+        if ($I->seePageHasElement($this->oldLoginSection)) {
+            $I->waitForElementVisible($this->userLoginEmail, 5);
+            $I->fillField($this->userLoginEmail, $userName);
+            $I->waitForElementVisible($this->userPassword, 5);
+            $I->fillField($this->userPassword, $userPassword);
+            $I->retryClick($this->loginButton);
+        }
+
+        if ($I->seePageHasElement($this->oneTouchNotNowLink)) {
+            $I->retryClick($this->oneTouchNotNowLink);
+        }
+
+        $this->waitForSpinnerDisappearance();
+        $this->removeCookieConsent();
+        $this->waitForSpinnerDisappearance();
+        $I->wait(3);
+    }
+
+   /**
+     * @param string $userName
+     * @param string $userPassword
+     */
+    public function approveGraphqlStandardPayPal(string $userName, string $userPassword): void
+    {
+        $I = $this->user;
+
+        $this->loginToPayPal($userName, $userPassword);
+
+        $this->confirmPayPal();
+
+        //retry
+        $this->waitForSpinnerDisappearance();
+        $this->confirmPayPal();
+
+        //we should be back to shop frontend as we sent a redirect url to paypal
+        $I->assertTrue($I->seePageHasElement($this->paypalBannerContainer));
+    }
+
+    public function confirmPayPal()
+    {
+        $I = $this->user;
+
+        $this->waitForSpinnerDisappearance();
+        $this->removeCookieConsent();
+
+        if ($I->seePageHasElement(substr($this->newConfirmButton, 1))) {
+            $I->retryClick($this->newConfirmButton);
+            $I->waitForDocumentReadyState();
+            $I->waitForElementNotVisible($this->globalSpinner, 60);
+            $I->wait(10);
+        }
+
+        if ($I->seePageHasElement("//input[@id='" . substr($this->newConfirmButton, 1) . "']")) {
+            $I->executeJS("document.getElementById('" . substr($this->newConfirmButton, 1) . "').click();");
+            $I->waitForDocumentReadyState();
+            $I->waitForElementNotVisible($this->globalSpinner, 60);
+            $I->wait(10);
+        }
+
+        if ($I->seePageHasElement($this->paymentConfirmButton)) {
+            $I->retryClick($this->paymentConfirmButton);
+            $I->waitForDocumentReadyState();
+            $I->waitForElementNotVisible($this->globalSpinner, 60);
+            $I->wait(10);
+        }
+    }
+
     public function waitForPayPalPage(): PayPalLogin
     {
         $I = $this->user;
@@ -152,5 +256,24 @@ class PayPalLogin extends Page
             $I->retryClick($this->acceptAllPaypalCookies);
             $I->waitForElementNotVisible($this->acceptAllPaypalCookies);
         }
-    }    
+    }
+
+    private function waitForSpinnerDisappearance()
+    {
+        $I = $this->user;
+        $I->waitForElementNotVisible($this->preloaderSpinner, 30);
+        $I->waitForElementNotVisible($this->globalSpinner, 30);
+        $I->waitForElementNotVisible($this->spinner, 30);
+    }
+
+    private function removeCookieConsent()
+    {
+        $I = $this->user;
+        if ($I->seePageHasElement($this->gdprContainer)) {
+            $I->executeJS("document.getElementById('" . substr($this->gdprContainer, 1) . "').remove();");
+        }
+        if ($I->seePageHasElement($this->gdprCookieBanner)) {
+            $I->executeJS("document.getElementById('" . substr($this->gdprCookieBanner, 1) . "').remove();");
+        }
+    }
 }