+++ CHANGELOG.md
@@ -12,6 +12,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Test compatibility with shop master.
 
+## [6.3.0] - Unreleased
+
+### Added
+- PayPal as payment method for OXID GraphQL Storefront module:
+  - services.yaml
+  - Code in namespace `\OxidEsales\PayPalModule\GraphQL\`
+  - Codeception tests in group `paypal_graphql`
+  - `\OxidEsales\PayPalModule\Model\PayPalRequest\GetExpressCheckoutDetailsRequestBuilder::setToken()`
+  - `\OxidEsales\PayPalModule\Model\PayPalRequest\GetExpressCheckoutDetailsRequestBuilder::getToken()`
+  - `\OxidEsales\PayPalModule\Model\PaymentManager`
+  - Column `oxuserbaskets.OEPAYPAL_PAYMENT_TOKEN` to hold the PayPal token needed for GraphQL checkout.
+
+### Changed
+- Changes visibility from protected to public for the following methods:
+   `\OxidEsales\PayPalModule\Model\Address::prepareDataPayPalAddress()`
+   `\OxidEsales\PayPalModule\Controller\StandardDispatcher::getReturnUrl()`
+   `\OxidEsales\PayPalModule\Controller\StandardDispatcher::getCancelUrl()`
+
+### Deprecated
+  - Deprecated the following methods:
+    `\OxidEsales\PayPalModule\Model\PayPalRequest\GetExpressCheckoutDetailsRequestBuilder::setSession()`
+    `\OxidEsales\PayPalModule\Model\PayPalRequest\GetExpressCheckoutDetailsRequestBuilder::getSession()`
+    `\OxidEsales\PayPalModule\Controller\StandardDispatcher::getTransactionMode()`
+
 ## [6.2.1] - 2020-10-08
 
 ### Fixed
@@ -203,6 +227,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Additional PayPal express checkout button in user checkout step in case no user is logged in.
 
 [7.0.0]: https://github.com/OXID-eSales/paypal/compare/b-6.x...master
+[6.3.0]: https://github.com/OXID-eSales/paypal/compare/v6.2.1...b-6.x
 [6.2.1]: https://github.com/OXID-eSales/paypal/compare/v6.2.0...v6.2.1
 [6.2.0]: https://github.com/OXID-eSales/paypal/compare/v6.1.0...v6.2.0
 [6.1.0]: https://github.com/OXID-eSales/paypal/compare/v6.0.2...v6.1.0